# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: kotoji Control-Plane API
  version: 1.0.0
  description: >
    The Go <-> Next REST/JSON control-plane API for kotoji. This spec is the SOURCE OF TRUTH
    for the Go <-> TS type sync: Go types are generated via oapi-codegen (-generate types) and
    the TS client via openapi-typescript + openapi-fetch. A CI drift gate fails on staleness.

    All schemas match the canonical Go core types frozen in CANONICAL.md. Wire field names are
    fixed there (e.g. baseSha, expected, actual, changedPaths). The data plane (static serving
    of *.hosting.example.com) is NOT part of this API and is specified in routing-and-serving.md.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: https://hosting.example.com
    description: Production control host (path-based split behind the proxy)
  - url: http://kotoji.localhost:8080
    description: Local dev
...
security:
  - sessionCookie: []
  - bearerPAT: []
paths:
...
  /api/sites/{handle}/publish:
    parameters:
      - $ref: "#/components/parameters/Handle"
    post:
      tags: [history]
      operationId: publish
      summary: Promote a source branch to published (direct mode) with optimistic locking
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublishRequest" }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublishResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Stale base SHA (conflict) or publish merge conflict (publish_conflict)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ConflictEnvelope"
                  - $ref: "#/components/schemas/PublishConflictEnvelope"
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-kotoji_session
      description: Opaque server-side session id (host-only __Host- cookie). Browser/dashboard auth.
    bearerPAT:
      type: http
      scheme: bearer
      bearerFormat: kotoji_pat
      description: >
        Per-site personal access token "kotoji_pat_<base62>". Scope is the intersection of the
        token scopes {read,write,publish} and the creating member's role (see CANONICAL.md §6).

  # --------------------------------------------------------------- parameters
  parameters:
    Handle:
      name: handle
      in: path
      required: true
      schema: { $ref: "#/components/schemas/HandleString" }
...
  responses:
...
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ValidationFailed:
      description: Validation failed (handle/path/branch/baseSha rules)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # --------------------------------------------------------------- schemas
  schemas:
    HandleString:
      type: string
      pattern: "^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$"
      minLength: 1
      maxLength: 63
      description: DNS-safe handle. Create-time min length is 3; resolver accepts 1-63. No "--".
      example: expense-calc
    BranchString:
      type: string
      pattern: "^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$"
      maxLength: 63
      description: Host-safe branch name. "published"/"draft" reserved; previews are feature-<user>-<slug>.
      example: draft
    Sha:
      type: string
      pattern: "^[0-9a-f]{40}([0-9a-f]{24})?$"
      description: Full git commit SHA (40 hex SHA-1 or 64 hex SHA-256).
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - validation
                - conflict
                - not_found
                - handle_taken
                - publish_conflict
                - branch_exists
                - nothing_to_commit
                - too_large
                - unsupported_media_type
                - rate_limited
                - quota_exceeded
                - internal
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              description: Code-specific structured detail (e.g. validation field, retry hint).
    ConflictEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string, enum: [conflict, handle_taken, branch_exists, nothing_to_commit] }
            message: { type: string }
            details: { $ref: "#/components/schemas/ConflictError" }
    ConflictError:
      type: object
      description: Optimistic-lock conflict detail (frozen field names; CANONICAL.md §8).
      required: [branch, expected, actual]
      properties:
        branch: { $ref: "#/components/schemas/BranchString" }
        expected: { type: string, description: "the baseSha the caller sent" }
        actual: { type: string, description: "the real current branch tip" }
        changedPaths:
          type: array
          items: { type: string }
    PublishConflictEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string, enum: [publish_conflict] }
            message: { type: string }
            details:
              type: object
              required: [paths]
              properties:
                paths:
                  type: array
                  items: { type: string }

    # ---------- auth ----------
...
    PublishRequest:
      type: object
      required: [baseSha]
      properties:
        from: { $ref: "#/components/schemas/BranchString", default: draft }
        baseSha: { type: string, description: "REQUIRED tip of `from` the publish is intended for" }
        message: { type: string }
    PublishResult:
      type: object
      required: [publishedCommit, from, fromCommit, pushed]
      properties:
        publishedCommit: { $ref: "#/components/schemas/Sha" }
        publishedUrl: { type: string, format: uri }
        from: { $ref: "#/components/schemas/BranchString" }
        fromCommit: { $ref: "#/components/schemas/Sha" }
        pushed: { type: boolean }
        warnings:
          type: array
          items: { type: string }
...